src/main/java/baseball/Utils/RandomNumberGenerator.java
package baseball.Utils;

import camp.nextstep.edu.missionutils.Randoms;

import java.util.ArrayList;
import java.util.List;

public class RandomNumberGenerator {
    private static final int MIN_NUMBER = 1;
    private static final int MAX_NUMBER = 9;
    private static final int NUMBER_OF_NUMBERS = 3;

    public static List<Integer> generate() {
        List<Integer> numbers = new ArrayList<>();

        while (numbers.size() < NUMBER_OF_NUMBERS) {
            int randomNumber = Randoms.pickNumberInRange(MIN_NUMBER, MAX_NUMBER);
            if (!numbers.contains(randomNumber)) {
                numbers.add(randomNumber);
            }
        }

        return numbers;
    }

}
